Add files via upload
in this project, first of all, understand the problem, then collect the data and clean it.
total casualties have taken place after the accident.
total casualties & percentage of total with respect to accident servity and maximum casualties by type of vehicle.
which vehicle has caused the most accident?
after that, it was found out that this year's and last year's monthly data were compared.
what was the type of Road that surfaced the most casualties? find out this from data.
what was the surface of the road so that it could be known what was the cause of the accident?
after that, it was to be found out that the accident had to be done to remove the relationship between the accident vs area .which area has the highest number of accidents?
and along with that, the area where the accident took place has happened during the day or at night.
in this project, I learned about pivot tables and pivot charts and how to use them. 
learn how to create a dashboard, learn about the calculation field, create KPI (key point indicator), and learn about top 10 analysis.